+++ lib/canopus/workspace/language_server_configurable.rb
@@ -98,6 +98,7 @@ module Canopus
         invalidate_document_links(client: client)
         invalidate_folding_ranges(client: client)
         invalidate_selection_ranges(client: client)
+        invalidate_hierarchy(client: client)
         invalidate_prepare_rename(client: client)
         invalidate_linked_editing_ranges(client: client)
         invalidate_inlay_hints(client: client)